LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/type.delete-expected.txt

type.delete
window.HTMLCanvasElement interface object is [[Configurable]]
Defined in "Web IDL" (draft)
Actual output:

PASS window.HTMLCanvasElement interface object is [[Configurable]]
